Created MVC setup, loading in stock, display stock
Created MVC file structure and created Stock controller and the User Views controllers to allow switching of the current view.
File reading in has been set up for the stock text file to create an initial stock level.
After reading in stock, items and levels are displayed.
Items can be added to the basket.