Admin login/panel improvements, fixing of adding items to baskets, general tweaks.
Admin login and admin panel now both have Controllers, as well as an Admin class for each admin's details to be stored as an object, which are then stored in the AdminDatabase to be checked upon login.

Changed error messages from console outputs to Message dialog boxes. Allows for clearer runtime messages of what has happened so that the user knows and allows for alerting of a system admin to what has actually gone wrong rather than it just crashing or catching an error and running incorrectly.

Items can now properly be added to baskets (see issues on github which is now closed for more details on what problem was and how it was rectified).

Otherwise, general tweaks were made to names etc. and comments added to methods as I went through to make it clearer for myself and others what they're doing (as vague method names can sometimes lead to mistakes about functionality).