Cash/Card payment added, receipt view setup, Payment controller set up
Cash payments fully working. Can enter a custom amount and the total to pay, how much has been paid, and the remainder to pay update each time an amount is entered. Once the total has been reached, a dialog box appears for the user to say if they want a receipt or not. If yes, opens the receipt view (currently not set up). If no, it returns the user to the till screen.

Card payment view has been set up, need to add functionality to accept & decline buttons.

Payment controller is used for both cash and card payments as it seems wasteful to create two seperate controllers for this

Minor changes include editing Product class to have a Product ArrayList, where the main Product type holds all of it's individual items in stock (as this is a use of the Composite design pattern). Identification of where the Facade pattern has been used (UserViewController being one) has been commented on, as well as other minor tweaks.