Card payment completed, receipt generation and threading completed, save stock to file implemented
Card payment screen has completed. If a bank card is declined, a dialog box will pop up asking whether they would like to try a different card or not. If yes, it simply returns to the Accept or Decline card screem. If no, it returns them to the kiosk and the basket is empty.

Receipt view has been set up. LocalThread is created in the ReceiptController. The BasketDatabase is then called to create the receipt. It creates a header (supermarket name, address, and the Date/Time of the transaction) and adds it to a DefaultModelList, it then goes through each item in the basket and gets the name and price each one and adds them individually to the list. Finally, it generates a footer (which is different whether they paid cash or card) to show how they paid, the total, and any change if they paid with cash. It then sets this into the thread and then into the JTextArea.

Save stock method has been created within StockController, currently being called when a Till View is generated (although this is just temporary, it will be changed to after orders are paid for and when deliveries are booked). Only the ArrayList of "barcodes" is not saved as these are generated based off the Product and the Product Code.

Otherwise, various tweaks and changes.