Output data fully working, duplicating data bug fixed, receipt format changed
The methd to save data is now working correctly. The error found with it repeatedly creating the list when saving was initially assumed to be as a result of the Product creation from inputting the Stock data (as a result of the composite pattern). However, it was coming from the same source as the ongoing bug I've had with the Available Stock display in the Till View displaying multiple copies of the data each time. Each time the loadGUI function was called within the UserViewController, the loadStock() method was running, causing it to read in more and more copies of the stock. To circumvent this, I've introducted an if-statement within the loadGUI method. It will only run the loadStock command if the StockDatabase is empty.

Receipt format changed to include the DecimalFormat of the popund currency "£0.00" so it doesn't end up with singular decimal points.