Admin product edit completed
Product Add/Edit has been created. This view contains the text fields and labels required for both creating a new Product and editing an existing one. It contains two seperate constructors which change certain things dependent on which one you want (for instance, when in Edit you get a combo box of all product to select which one to change, which isn't needed). Update fully cascades and edits existing products as well as creating new products for the updated item to be stored within the Barcodes ArrayList.