Core functionality completed, invoices can be generated,  sound effects added, minor tweaks made.
Kiosk now has full functionality. All required operations have been completed and the till is fully operational.
Admin panel users can now generate invoices. The file name is set to the current date/time of when the invoice is ran, it contains a header and footer and lists the products required and the product cost per unit. The invoice automatically orders whatever the minimum stock level is * 2, but only for products that are below their minimum stock levels. The invoice also contains a footer with the invoice total for goods.
The update product allows for entering of received deliveries.

To show the kiosk system off in a more realistic fashion, I have added some sound effects to simulate what you would hear when in use. The scan barcode button now randomly picks one of the Products and adds it to basket (in real use, this would just function as an actual barcode scanner). When the scan barcode, or manually adding an item, you will hear a barcode scanner beeping. If you choose to have a receipt print out, you will hear the sound of a receipt printing and being ripped.

Finally, various tweaks and changes have been made just for the purpose of refining.

The next stages are to: remove redundant code, refine any methods I can, implement tests to ensure the expected outcome is also correct, make changes where possible to incorporate more design patterns, and improve the GUI.